feat/update: matrícula normal e matrícula extraordinária; alteração em models; alteração da representação de coleções (turmas_UnB -> turmas);

+++ backend/models/turma.go
@@ -9,12 +9,12 @@ type Turma struct {
 	MateriaId       string            `json:"materiaId" firestore:"materiaId"`
 	NomeMateria     string            `json:"nomeMateria" firestore:"nomeMateria"`
 	Semestre        string            `json:"semestre" firestore:"semestre"`
-	VagasTotais     int               `json:"vagasTotais" firestore:"vagasTotais"`
-	VagasOcupadas   int               `json:"vagasOcupadas" firestore:"vagasOcupadas"`
+	VagasTotais     int64               `json:"vagasTotais" firestore:"vagasTotais"`
+	VagasOcupadas   int64               `json:"vagasOcupadas" firestore:"vagasOcupadas"`
 	Ativo           bool              `json:"ativo" firestore:"ativo"`
 	Local           string            `json:"local" firestore:"local"`
 	Horario         string            `json:"horario" firestore:"horario"`
-	VagasExclusivas map[string]int    `json:"vagasExclusivas" firestore:"vagasExclusivas"`
+	VagasExclusivas map[string]int64    `json:"vagasExclusivas" firestore:"vagasExclusivas"`
 	ProfessorId     string            `json:"professorId" firestore:"professorid"`
 	ProfessorNome   string            `json:"professorNome" firestore:"professorNome"`
 	Prioridades     map[string]string `json:"prioridades" firestore:"prioridades"`
